+++ terraform/edge-network/variables.tf
@@ -26,16 +26,18 @@ variable "network_bridge" {
   default     = "virbr1"
 }
 
-variable "network_domain" {
-  type        = string
-  description = "The domain of the network"
-  default     = "cluster.onprem"
-}
-
 variable "network_subnet_cidrs" {
   type        = list(string)
   description = "The CIDRs of the network"
   default     = ["192.168.99.0/24"]
+}
+
+variable "dns_domain" {
+  type        = string
+  description = "The DNS domain"
+  # TODO: Change this to the correct domain name
+  # default     = "cluster.onprem"
+  default     = "demo.onprem.espdqa.infra-host.com"
 }
 
 variable "dns_hosts" {
@@ -45,46 +47,39 @@ variable "dns_hosts" {
   }))
   description = "The list of DNS hosts"
   default = [
-    { hostname = "alerting-monitor.cluster.onprem", ip = "192.168.99.30" },
-    { hostname = "api.cluster.onprem", ip = "192.168.99.30" },
-    { hostname = "app-orch.cluster.onprem", ip = "192.168.99.30" },
-    { hostname = "app-service-proxy.cluster.onprem", ip = "192.168.99.30" },
-    { hostname = "argocd.cluster.onprem", ip = "192.168.99.20" },
-    { hostname = "attest-node.cluster.onprem", ip = "192.168.99.30" },
-    { hostname = "cluster-orch-edge-node.cluster.onprem", ip = "192.168.99.30" },
-    { hostname = "cluster-orch-node.cluster.onprem", ip = "192.168.99.30" },
-    { hostname = "cluster.onprem", ip = "192.168.99.30" },
-    { hostname = "connect-gateway.cluster.onprem", ip = "192.168.99.30" },
-    { hostname = "fleet.cluster.onprem", ip = "192.168.99.30" },
-    { hostname = "infra-node.cluster.onprem", ip = "192.168.99.30" },
-    { hostname = "keycloak.cluster.onprem", ip = "192.168.99.30" },
-    { hostname = "license-node.cluster.onprem", ip = "192.168.99.30" },
-    { hostname = "log-query.cluster.onprem", ip = "192.168.99.30" },
-    { hostname = "logs-node.cluster.onprem", ip = "192.168.99.30" },
-    { hostname = "metadata.cluster.onprem", ip = "192.168.99.30" },
-    { hostname = "metrics-node.cluster.onprem", ip = "192.168.99.30" },
-    { hostname = "observability-admin.cluster.onprem", ip = "192.168.99.30" },
-    { hostname = "observability-ui.cluster.onprem", ip = "192.168.99.30" },
-    { hostname = "onboarding-node.cluster.onprem", ip = "192.168.99.30" },
-    { hostname = "onboarding-stream.cluster.onprem", ip = "192.168.99.30" },
-    { hostname = "registry-oci.cluster.onprem", ip = "192.168.99.30" },
-    { hostname = "registry.cluster.onprem", ip = "192.168.99.30" },
-    { hostname = "release.cluster.onprem", ip = "192.168.99.30" },
-    { hostname = "telemetry-node.cluster.onprem", ip = "192.168.99.30" },
-    { hostname = "tinkerbell-nginx.cluster.onprem", ip = "192.168.99.40" },
-    { hostname = "tinkerbell-server.cluster.onprem", ip = "192.168.99.30" },
-    { hostname = "update-node.cluster.onprem", ip = "192.168.99.30" },
-    { hostname = "vault.cluster.onprem", ip = "192.168.99.30" },
-    { hostname = "vnc.cluster.onprem", ip = "192.168.99.30" },
-    { hostname = "web-ui.cluster.onprem", ip = "192.168.99.30" },
-    { hostname = "ws-app-service-proxy.cluster.onprem", ip = "192.168.99.30" },
+    { hostname = "alerting-monitor", ip = "192.168.99.30" },
+    { hostname = "api", ip = "192.168.99.30" },
+    { hostname = "app-orch", ip = "192.168.99.30" },
+    { hostname = "app-service-proxy", ip = "192.168.99.30" },
+    { hostname = "argocd", ip = "192.168.99.20" },
+    { hostname = "attest-node", ip = "192.168.99.30" },
+    { hostname = "cluster-orch-edge-node", ip = "192.168.99.30" },
+    { hostname = "cluster-orch-node", ip = "192.168.99.30" },
+    { hostname = "connect-gateway", ip = "192.168.99.30" },
+    { hostname = "fleet", ip = "192.168.99.30" },
+    { hostname = "infra-node", ip = "192.168.99.30" },
+    { hostname = "keycloak", ip = "192.168.99.30" },
+    { hostname = "license-node", ip = "192.168.99.30" },
+    { hostname = "log-query", ip = "192.168.99.30" },
+    { hostname = "logs-node", ip = "192.168.99.30" },
+    { hostname = "metadata", ip = "192.168.99.30" },
+    { hostname = "metrics-node", ip = "192.168.99.30" },
+    { hostname = "observability-admin", ip = "192.168.99.30" },
+    { hostname = "observability-ui", ip = "192.168.99.30" },
+    { hostname = "onboarding-node", ip = "192.168.99.30" },
+    { hostname = "onboarding-stream", ip = "192.168.99.30" },
+    { hostname = "registry-oci", ip = "192.168.99.30" },
+    { hostname = "registry", ip = "192.168.99.30" },
+    { hostname = "release", ip = "192.168.99.30" },
+    { hostname = "telemetry-node", ip = "192.168.99.30" },
+    { hostname = "tinkerbell-nginx", ip = "192.168.99.40" },
+    { hostname = "tinkerbell-server", ip = "192.168.99.30" },
+    { hostname = "update-node", ip = "192.168.99.30" },
+    { hostname = "vault", ip = "192.168.99.30" },
+    { hostname = "vnc", ip = "192.168.99.30" },
+    { hostname = "web-ui", ip = "192.168.99.30" },
+    { hostname = "ws-app-service-proxy", ip = "192.168.99.30" },
   ]
-}
-
-variable "dhcp_boot" {
-  type        = string
-  description = "The DHCP boot option"
-  default     = "tag:efi-http,https://tinkerbell-nginx.cluster.onprem/tink-stack/signed_ipxe.efi"
 }
 
 variable "dns_resolvers" {
